Added heart icons
Added the hearts from last year into the assets folder. They don't line up with Andrew's health system, so I just grabbed a random heart png from the internet, scaled it appropriately, and added it to the health system, so the hearts are actual hearts and not the godot logo.