Started created a separate HUD scene
I think it would be nice to move the health UI elements over to a new scene so it's easier to work with. I have created a new scene, and added 3 hearts and a spear that will hover in the top left of the game screen. I will work on moving Andrew's functionality over to it next.