connecting signals in-script
Tried to connect health_changed signal to a function in levelRoot via script. This should happen after the player scene is instanced and added as a child. It does not, apparently. The signal is "nonexistent" per the error (replicate by running as is and opening the debug level, see debugger window)