interactBase is now ready for doing stuff!
Important notes beyond interactables: The pause modes for debug level's ambience sound and the player footsteps are now "process". The pause state handling has changed fundamentally.

When you press E in the presence (Area2D) of the fingers, "Elden John" pops up and the game_tree pauses.

The pause state was modified so you can't accidentally pause the game during an interaction.

Right now to exit the interaction you press E again.

Look in rootLevel's "_input(_event) function for pause screen changes (in short, checks if the game is paused while the pauseScreen is not showing).

Also, to make it less jarring, I set the player's foot steps and the level's ambience AudioPlayerModes to "process" in their pause modes.