Completed basis of level transition
Instances of the elevatorExit now allow you to move between levels with a brief menu. For the sake of demonstration, the debugLevel has a sequel with a (really stupid) surprise. The elevator does not have an animation yet, unfortunately.

The new debugLevelTwo is a basic arena where the player gets dumped into

the levelRoot node has a new function that connects to a signal in the elevator scene (called levelTransition)

The player's collision layer bits get  set in its _ready function and processState() function (will be investigated later)

New single frame AnimatedSprite resource for the elevator was created (will use this for later elevator stuff)

Will probably make a new commit that rearranges the prompts, as they currently have poor visibility