More Red Boss stuff (hands)
Added hands that are supposed to stay within a specific range of the boss but attack the player when they are in range of player which the player can destroy and some tweaks for stomps and stun mechanics. Obviously, still far from done as the boss room has just been completed and still needs animations.