boss and pegasus bug fixes
This should fix most of the bugs of the boss and the pegasus enemy. there are still some bugs but play testing should give an idea of what really needs to be fixed. also updated the boss sprites, they still need work and need to add sounds but will stay the way it for purely play testing.